Moved address pool-related code into dhcp6/pool package.
