+++ dhcp6/pool/random_address_pool_test.go
@@ -1,4 +1,4 @@
-package dhcp6
+package pool
 
 import (
 	"testing"
